Character model is now soft-deleteable
Implemented a SoftDeleteable pattern for the character model that can be
freely adapted to other models as well.

In contrary to the SoftDeleteable-Extension available, this pattern only
filters if the results if the repository-methods find and findAll are
called.

Tests where added to test for this behaviour. MotD now returns a
soft-deleted character to be able to access its properties.