Extends the message model with tests
The character model now supports the listing of received and sent messages
if the message is persisted explicitely.

For use with the Creator-Trait, a new interface CreateableInterface has
been introduced that explicitly type hints for the interface. The trait
has been changed accordingly.

Re-introduces the MissingCharacter which now defaults to "Nobody" if the
character really disappeared from the database.