Changed the MessageModel as discussed
The message model now supports only a single message line without any
title. Furthermore, it has an author as well as a thread_id.

The thread_id links to a table with threads which are collections of
messages that belong to certain characters. Using this approach, the
model supports One-to-One-Chat, Group-Chat, System-to-One-Chat, as well as
a System-to-Many-Chat.

Using this model, commentary could be implemented using this entity as
well.